Fix pre-commit failures
