+++ build_tools/github_actions/test_reporting_script.py
@@ -51,43 +51,42 @@ def create_index_file(args: argparse.Namespace):
     build_dir = args.build_dir / "artifacts"
 
     indexer_args = argparse.Namespace()
-    indexer_args.filter = ["*.tar.xz*"]
+    indexer_args.filter = ["*.html*"]
     indexer_args.output_file = "index.html"
     indexer_args.verbose = False
     indexer_args.recursive = False
     process_dir(build_dir, indexer_args)
 
 
-def upload_artifacts(args: argparse.Namespace, bucket_uri: str):
-    logging.info("Uploading artifacts to S3")
-    build_dir = args.build_dir
-    amdgpu_family = args.amdgpu_family
+# Enhancement to upload all HTML test reports in a directory
+def upload_test_report(report_dir: Path, bucket_uri: str):
+    """
+    Upload all .html files from report_dir to bucket_uri (keeps filenames).
+    """
+    if not report_dir.exists() or not report_dir.is_dir():
+        logging.error(
+            "Report directory %s not found or not a directory — skipping upload.",
+            report_dir,
+        )
+        return
 
-    # Uploading artifacts to S3 bucket
+    # Use a single AWS CLI call to copy only *.html files recursively
     cmd = [
         "aws",
         "s3",
         "cp",
-        str(build_dir / "artifacts"),
+        str(report_dir),
         bucket_uri,
         "--recursive",
-        "--no-follow-symlinks",
         "--exclude",
         "*",
         "--include",
-        "*.tar.xz*",
+        "*.html",
+        "--content-type",
+        "text/html",
     ]
     exec(cmd, cwd=Path.cwd())
-
-    # Uploading index.html to S3 bucket
-    cmd = [
-        "aws",
-        "s3",
-        "cp",
-        str(build_dir / "artifacts" / "index.html"),
-        f"{bucket_uri}/index-{amdgpu_family}.html",
-    ]
-    exec(cmd, cwd=Path.cwd())
+    logging.info("Uploaded all .html files from %s to %s", report_dir, bucket_uri)
 
 
 def run(args: argparse.Namespace):
@@ -95,8 +94,22 @@ def run(args: argparse.Namespace):
     run_id = args.run_id
     bucket_uri = f"s3://{bucket}/{external_repo_path}{run_id}-{PLATFORM}"
 
+    # Skip uploading report if report path is not set
+    if args.report_path is None:
+        logging.error("--report-path is not provided — skipping upload")
+        return
+    if not args.report_path.exists():
+        logging.error(
+            "--report-path %s does not exist — skipping upload", args.report_path
+        )
+        return
+    logging.info(
+        "--report-path is set; uploading HTML reports from %s to %s",
+        args.report_path,
+        bucket_uri,
+    )
+    upload_test_report(args.report_path, bucket_uri)
     create_index_file(args)
-    upload_artifacts(args, bucket_uri)
 
 
 def main(argv):
@@ -114,6 +127,12 @@ def main(argv):
         type=Path,
         required=True,
         help="Path to the build directory of TheRock",
+    )
+
+    parser.add_argument(
+        "--report-path",
+        type=Path,
+        help="Directory containing .html files to upload (optional)",
     )
 
     args = parser.parse_args(argv)
